Commiting a literate programming approach to hplc analysis with org
Literate programming is a powerful tool, which is very usable inside
org mode. This org document contains all the code written for HPLC
data analysis so far and tangles it to the correct scripts, as well as
weaves it to a pdf file that can be written outside of emacs. It is a
very powerful tool and one I wanted to leverage for this repository