Adding the last bunch of pre-existing info to the notebook.
The metabolic pathways notebooks which has been fleshed out in the
last few commits contains a more organized version of the julia files
for the metabolic pathways project I have been working on. The project
is nowhere near complete but an important milestone of writing down
all the core pathways which we expect was achieved. In next commits,
it will be more fleshed out with things like only taking sugars
that actually went to products in mind (because sugars can also be
oxidized by the oxygen that is periodically in the reactor when we
take samples, can be consumed for the production of microbial biomass
or even can be consumed and stuck on an intermediate of the pathway
which we don't track, which is the main problem of modeling a
transient state with steady state equations). Another interesting
concept that is being considered is including these things for the
products as well, because in the kinetic experiment, there was a
noticeable decrease in products which could not be described as them
going to a different product in some cases.