Adding a notebook and supporting files for hypothesis testing.
The goal of this notebook is a step by step explanation of the ANOVA
hypothesis test done for my thesis data to show that the parameters we
are changing significantly affect the process. Besides the original
notebook and its exported version, in this commit there are some csv
files, which contain results of the hypothesis tests done and the
tangled code files. One contains the preprocessing of the data and
creating distributions from which we can sample data to do the
analysis and the other has the code of the hypothesis tests.