Adding some useful processed data and plots.
The csv files contain the products of the experiment in 45 C separated
from the rest. This is useful cause we plan to use it in the
hypothesis testing file. The pngs are scatter plots for visualizing
the product concentrations in each temperature and mix amount. This is
important to show as it is the basis of hypothesis testing and
sensitivity analysis.