A lot more work on the metabolic pathways project.
The system is now trained for the ethanol and acetate before
acetogenic reactions and inserts them after. This causes problems in
the three reactions where acetogenesis is completely inhibited, but
for the other 7 is a necessary part. Also to accomodate for this
change, the signature of the generic functions now uses the dataframe
and not a final state.

Error handling was also added to the optimization interface's loss
function as during optimization we don't care if we step on errors as
long as we move away from them (hence why the value the loss returns
if it hits an error is large).

With these tools, the other 9 reactions were optimized and we now have
a vector of parameters for each of the reactions, except for 35_2,
which seems to be stuck in a local minima and I can't find a way to
make it converge at least for now.