More work on the metabolic pathways project.
In the previous commit, I worked on the optimization problems and
finding the ideal parameters for each system to be described. In this
commit, I saved those results in a jld2 file in the simulations folder
to be read from other files without needing to re-run the
optimizations and then did a lot of data processing on that to get to
the point where I can generate tables with all the metabolic fluxes
and a graph visualizing them.