Adding a manuscripts folder to the directory.
In this folder, all initial manuscripts of my thesis will be
included. In this commit, is a handwritten manuscript from xournal++,
the beginning of the introduction manuscripts, containing why food
waste is a problem, why enzymatic hydrolysis is the best pretreatment
and why it is not economically feasible (some of which will also be
augmented) and a file containing a very short description of each
article I have so far read and plan to use in my thesis.