Commiting a new notebook for data processing of anaerobic digestion.
This notebook will be used for analyzing all AD experiments. Some of
the code blocks are written in such a way, so that they can be re-used
for every experiment very easily. Then, the rest of the code blocks
apply that to each experiment, making this a very reproducible system
for analyzing these BMP results.