Adding an org document with a statistics introduction section.
Since we had access to Statista for statistical data on various
sectors, I had collected a list of useful plots concerning food waste,
waste treatment, energy production and bioenergy. All these are going
to be summarized in this manuscript, which is going to be at the very
start of the thesis.