Adding another useful to the AD analysis notebook.
This commit adds a new useful plot for AD analysis. The idea is to
compare the volume of methane produced from acetate with that produced
from the hydrolysates. For that, the final methane volume of each
experiment was added to its evaluation (which was added from the
previous commit to accomodate these changes) and then a plot was made
visualizing these with the CairoMakie.jl package.