Changes in the AD notebook following the previous commits.
In previous commits, various tasks within the AD notebook were
oultined. This commit implements them in the notebook.

Furthermore, another cool feature was added to the notebook to use
auto updating org-tables for all the data I want to input. This is
done by updating the table by re-reading the CSV and inserting it,
using some org-mode magic.