More work on the COD notebook for hydrolysis data.
One of the most common ways in literature to assess the quality of
acidification in 2-phase AD is to measure the acidification rate which
is the tVFAs expressed in COD-eq divided by the sCOD (this shows how
much of the COD is acidogenic products). This commit does these kind
of calculations and places them in the COD notebook that was created
in the past.