Adding some useful hydrolysis plots.
A plot type that I have seen a lot in literature is a stacked bar plot
showing the amount of acidogenic products during different phases of a
2-stage AD process. I took this idea and made such plots to compare
our batch experiments. Such a plot was used as the introduction to the
notebook for hypothesis testing and sensitivity, so that code block
was changed to make these kind of plots.