Major change to the plotting of the AD analysis.
The plots using ml CH4/g sCOD are removed and in their place ml CH4/g
VS are added. This change was made because to represent it on a COD
basis, you need to use the consumed COD and not the added one and the
former is very hard to measure. The VS basis is the VS of the sludge
added to the reactor which is known. Also, the kinetics were fit on
both ml CH4 and ml CH4/g VS.