Exported versions of the AD presentation from the previous commit.
In these exports there is an html file, which is a reveal.js
presentation exported with org-re-reveal as well as the according pdf
file exported using the decktape utility. This is the finalized
presentation format that I will use. In the commit is also a ppt and
exported pdf version of the presentation. This was the one made
initially, but after that, I tried doing it in reveal and it turned
out better. This was mainly done because I found out about decktape to
finally have a working pdf version of the presentation.