Correcting some data for the S2_R1 experiment.
There was some inaccuracy in this experiment with the methane
production potential being very high. The correction that was done was
to subtract these measurements based on how they were found during the
acetate experiment from the hydrolysate experiments, which yields much
more logical results. This is yet to be documented on the notebook (as
with many other things in it).