Organizing the AD notebook further.
This commit concerns a lot of organizational things about the AD
notebook I have developed in previous commits. First, it documents the
hydrolysate experiment with the second sludge much better. Secondly,
it restructures how I do tangling to make for a much cleaner result
which is easier for someone to read.